Update 1.0.1, change of lots of stuff!
This update fixes lots of little bugs running out there.
The main purpose for this was to make it able to replace the images even after the page loaded! So, if the page you're looking loads images with JavaScript after the page loads, the extension will replace them! It'll try to find new images for 10 times with 1.5 second delay. When it couldn't find any new imgur images in 15 seconds, it'll stop checking!
And also auto activation of the extension is working now. _It wasn't before, yeah you had to click to popup or visit an imgur image for the first time to get it activated._ But it works now! Wooho!